Fixes some image patches bugs

+++ ummon/utils.py
@@ -174,7 +174,7 @@ def gini(array):
     # from:
     # http://www.statsdirect.com/help/default.htm#nonparametric_methods/gini.htm
     # All values are treated equally, arrays must be 1d:
-    array = array.flatten()
+    array = array.flatten().astype(np.float32)
     if np.amin(array) < 0:
         # Values cannot be negative:
         array -= np.amin(array)
